+++ messaging/src/main/java/de/fraunhofer/ids/messaging/protocol/http/HttpService.java
@@ -6,7 +6,7 @@ import java.time.Duration;
 import java.util.Map;
 
 import de.fraunhofer.ids.messaging.core.daps.ClaimsException;
-import de.fraunhofer.ids.messaging.core.util.MultipartParseException;
+import de.fraunhofer.ids.messaging.protocol.multipart.parser.MultipartParseException;
 import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
