+++ messaging/src/main/java/de/fraunhofer/ids/messaging/protocol/http/IdsHttpService.java
@@ -17,9 +17,9 @@ import de.fraunhofer.ids.messaging.core.config.ClientProvider;
 import de.fraunhofer.ids.messaging.core.config.ConfigContainer;
 import de.fraunhofer.ids.messaging.core.daps.ClaimsException;
 import de.fraunhofer.ids.messaging.core.daps.DapsValidator;
-import de.fraunhofer.ids.messaging.core.util.MultipartDatapart;
-import de.fraunhofer.ids.messaging.core.util.MultipartParseException;
-import de.fraunhofer.ids.messaging.core.util.MultipartParser;
+import de.fraunhofer.ids.messaging.protocol.multipart.parser.MultipartDatapart;
+import de.fraunhofer.ids.messaging.protocol.multipart.parser.MultipartParseException;
+import de.fraunhofer.ids.messaging.protocol.multipart.parser.MultipartParser;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
