+++ messaging/src/main/java/de/fraunhofer/ids/messaging/protocol/multipart/parser/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Utility classes, for parsing MultiPart Messages and enums.
  */
-package de.fraunhofer.ids.messaging.core.util;
+package de.fraunhofer.ids.messaging.protocol.multipart.parser;
