+++ messaging/src/main/java/de/fraunhofer/ids/messaging/util/InfomodelMessageBuilder.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 
 import de.fraunhofer.iais.eis.Message;
 import de.fraunhofer.iais.eis.ids.jsonld.Serializer;
-import de.fraunhofer.ids.messaging.core.util.MultipartDatapart;
+import de.fraunhofer.ids.messaging.protocol.multipart.parser.MultipartDatapart;
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
 import okhttp3.MediaType;
